ctf-server/pyproject.toml
[tool.poetry]
name = "ctf-server"
version = "0.1.0"
description = ""
authors = ["memoey <git@memoey.icu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
astroid = "3.1.0"
build = "1.1.1"
cachecontrol = "0.14.0"
certifi = "2024.2.2"
cffi = "1.16.0"
cfgv = "3.4.0"
charset-normalizer = "3.3.2"
cleo = "2.1.0"
crashtest = "0.4.1"
cryptography = "42.0.5"
dill = "0.3.8"
distlib = "0.3.8"
dulwich = "0.21.7"
fastjsonschema = "2.19.1"
filelock = "3.13.1"
identify = "2.5.35"
idna = "3.6"
importlib-metadata = "7.0.2"
iniconfig = "2.0.0"
installer = "0.7.0"
isort = "5.13.2"
jaraco-classes = "3.3.1"
jeepney = "0.8.0"
keyring = "24.3.1"
mccabe = "0.7.0"
more-itertools = "10.2.0"
msgpack = "1.0.8"
nodeenv = "1.8.0"
packaging = "23.2"
pexpect = "4.9.0"
pkginfo = "1.10.0"
platformdirs = "4.2.0"
pluggy = "1.4.0"
poetry = "1.8.2"
poetry-core = "1.9.0"
poetry-plugin-export = "1.6.0"
pre-commit = "3.6.2"
ptyprocess = "0.7.0"
pycparser = "2.21"
pylint = "3.1.0"
pyproject-hooks = "1.0.0"
pytest = "8.0.2"
pyyaml = "6.0.1"
rapidfuzz = "3.6.2"
requests = "2.31.0"
requests-toolbelt = "1.0.0"
secretstorage = "3.3.3"
shellingham = "1.5.4"
tomlkit = "0.12.4"
trove-classifiers = "2024.3.3"
urllib3 = "2.2.1"
virtualenv = "20.25.1"
zipp = "3.17.0"
pydantic = "^2.6.3"
uvicorn = "^0.27.1"
fastapi = "^0.110.0"
chardet = "5.1.0"
click = "8.1.3"
colorama = "0.4.6"
pip-upgrade = "0.0.6"
pur = "7.1.0"
six = "1.16.0"
azure-core = "^1.30.1"
azure-cosmos = "^4.5.1"
pytest-cov = "^4.1.0"
testcontainers-mongodb = "^0.0.1rc1"
pymongo = "^4.6.2"
azure-functions = "^1.18.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
